ten_ninety
New program to calculate the percentage-change over the last year, and
then print out the 10th and 90th percentile of that.  This makes it
easier to find out when to even consider selling options.

Selling options of existing stock should be done when the perc-change
is above the 90th-percentile, and we should purchase-more-share and
then get covered-calls when the stock price goes below the 10-mark.