NN in JS
Now porting the code from C to JS, because I'd like to see how this
fares.  I will be keeping the code in a C-like shape, since I enjoy
the simplicity that comes from imperative programming.  It'll be
pretty nice, though, to be able to write essentially C code with a
bunch of conveniences from a higher-level language.